Adding Hyperload bootloader as top level project
- Added Hyperload Bootloader as an example project
- Added new make target, "make bootloader" and "make application"
- make bootloader will place the IVT at address 0 in flash and Disabled FreeRTOS
- make application will build project with IVT at 64kB from start of flash and
  links FreeRTOS
- ResetISR now manually moves stack pointers back to proper position after
  bootloader has run.
- Gave uart0::PutChar and uart0::GetChar necessary timeouts to work with the
  Hyperload protocol.
- Slight modification to SystemTimer where all control registers are zeroed when
  disabled
- Added linker script for application and bootloader
- Added hyperload bootloader building to presubmit checks
- Added some initial testing of uart0.cpp

+++ firmware/library/L2_Utilities/debug.hpp
@@ -1,8 +1,9 @@
 #pragma once
 
+#include <cstdint>
 #include <cstdlib>
 
 namespace debug
 {
-void Hexdump(void * address, size_t length);
+void Hexdump(void * address, uint32_t length);
 }  // namespace debug
